refactor: better typing for reducers & readme logo update
